requirements.txt
pandas>=1.5.0
scikit-learn>=1.0.0
numpy>=1.21.0
pyyaml>=6.0
jupyter>=1.0.0
transformers>=4.0.0
datasets>=2.0.0
torch>=1.10.0
joblib>=1.1.0 